Adding a prototype that has a copy button for the decryption input box
This is technically adding a feature, and I don't want to canonize it because it looks bad on the computer I used to make it. If I'm showing this program in a job interview or other IRL demo, then I want the download link to visibly work on the live page. The computer I'm doing this on is *not* a computer that is for frivolity. Also it's 16:10...

For my own job security, I will not be canonizing this prototype due to it bugging out (sometimes even relocating page elements, and even if the UI doesn't suck, I can't risk it, even though I haven't reproduced the bug again, I can't afford falling flat on my face in THAT way again, so no dice), and its UI being difficult to make look good. Also, I will not be updating Flashpoint's at all because the 2024 file loaded in Flashpoint has problems. And yes, different Firefox forks handle it differently. Not something I want to deal with.

Also, not just in-person demos either tbh. And yes, I will use the http://b3k.sourceforge.io link due to it being simpler.

I do hope to be employed sooner rather than later, and this computer is my most portable non-obsolete one. Also I'm taking some IT courses in college and I at least want to be able to show people the program working rather than them seeing no DL link. Also, if I'm going to demonstrate the program on YouTube with screen sharing, or livestream using it, I can't have an invisible download link otherwise people will complain. Same for two copy buttons. I can already see the contention on that. Not something I want to deal with. Did I mention there's no room to not make the buttons confusing on MY PC. Also, two copy buttons would confuse people, AND each element takes up more browser memory.

Use this as you please on your own computers, but until I find a solution, I ain't putting this on GH pages or SourceForge at all. I'll be sure and let y'xll know when I do.